Fixed the way of importing the moved model classes
